pyproject.toml
[tool.poetry]
name = "ska-ser-infra-machinery"
version = "0.0.1"
description = ""
authors = ["Domingos Nunes <domingos.nunes@atlar.pt>"]

[tool.poetry.dependencies]
python = "^3.8"
ansible = "^6.5.0"
python-openstackclient = "^6.0.0"
hvac = "^1.0.2"


[tool.poetry.dev-dependencies]
ansible-lint = "^6.8.6"
yamllint = "^1.28.0"
ansible-lint-junit = "^0.17.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
